requirements.txt
google.genai
requests
python-dotenv
pymongo
Flask
flask-cors
Flask-Bcrypt